requirements.txt
fastapi
uvicorn
python-multipart
jinja2
python-dotenv
requests
opencv-python
pillow
numpy
pandas
imagehash
rawpy
sqlalchemy
pyinstaller
pywebview
openai
